Update after code review
