+++ requirements.txt
@@ -5,3 +5,4 @@ python-keystoneclient>=0.10.0,<=1.1.0
 PyYAML>=3.1.0,<=3.10
 requests>=2.1.0,!=2.4.0,<=2.2.1
 six>=1.7.0,<=1.9.0
+netaddr>=0.7.14
